requirements.txt
aiohttp==3.9.5
httpx==1.8.3
python-dotenv==1.0.0
jinja2==3.1.2
